# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Maestro Symphony",
    "description": "Maestro Symphony is a fast, mempool-aware Bitcoin indexer and API server. It supports mainnet and testnet4, providing data for UTXOs, metaprotocols and much more. Designed for developers who need reliable, mempool-aware blockchain data.",
    "contact": {
      "name": "jharper@gomaestro.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "v0.0.1"
  },
  "paths": {
    "/addresses/{address}/runes/balances": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Rune Balances by Address",
        "description": "Returns total rune balances held in UTxOs controlled by the provided address, sorted by rune ID.",
        "operationId": "addresses_all_rune_balances",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tb1qphcdyah2e4vtpxn56hsz3p6kapg90pl4x525kc"
          },
          {
            "name": "mempool",
            "in": "query",
            "description": "Mempool-aware (default: false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeResponse_Vec_RuneAndAmount"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/addresses/{address}/runes/balances/{rune}": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Rune Balance by Address and Rune",
        "description": "Returns the total amount of the provided rune kind held in UTxOs controlled by the provided address.",
        "operationId": "addresses_specific_rune_balance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tb1qphcdyah2e4vtpxn56hsz3p6kapg90pl4x525kc"
          },
          {
            "name": "rune",
            "in": "path",
            "description": "Rune ID or name (spaced or unspaced)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "BITCOIN•PIZZAS"
          },
          {
            "name": "mempool",
            "in": "query",
            "description": "Mempool-aware",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeResponse_RuneAndAmount"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/addresses/{address}/runes/tx/{txid}": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Rune Edicts by Address and Transaction",
        "operationId": "addresses_runes_by_tx",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tb1qphcdyah2e4vtpxn56hsz3p6kapg90pl4x525kc"
          },
          {
            "name": "txid",
            "in": "path",
            "description": "Transaction ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "c0345bb5906257a05cdc2d11b6580ce75fdfe8b7ac09b7b2711d435e2ba0a9b3"
          },
          {
            "name": "mempool",
            "in": "query",
            "description": "Mempool-aware",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeResponse_Vec_RuneEdict"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/addresses/{address}/runes/utxos": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Rune UTxOs by Address",
        "description": "Returns all UTxOs controlled by the provided address which contain runes, sorted by height.",
        "operationId": "addresses_all_rune_utxos",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tb1qphcdyah2e4vtpxn56hsz3p6kapg90pl4x525kc"
          },
          {
            "name": "mempool",
            "in": "query",
            "description": "Mempool-aware (default: false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServeResponse_Vec_RuneUtxo"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChainTip": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string",
            "description": "The hash of the block"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "description": "The height of the block in the blockchain",
            "minimum": 0
          }
        }
      },
      "EstimatedBlock": {
        "type": "object",
        "required": [
          "block_height"
        ],
        "properties": {
          "block_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "IndexerInfo": {
        "type": "object",
        "required": [
          "chain_tip",
          "estimated_blocks"
        ],
        "properties": {
          "chain_tip": {
            "$ref": "#/components/schemas/ChainTip",
            "description": "Most recent *mined* block in the indexed mainchain, any estimated blocks will be descendants of this block"
          },
          "estimated_blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EstimatedBlock"
            },
            "description": "Information about any estimated blocks from the mempool that were indexed in addition to the mainchain"
          },
          "mempool_timestamp": {
            "type": [
              "string",
              "null"
            ],
            "description": "Timestamp of the indexed mempool snapshot, if any estimated blocks from the mempool have been indexed"
          }
        }
      },
...
      "RuneAndAmount": {
        "type": "object",
        "required": [
          "id",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
...
      "ServeResponse_RuneAndAmount": {
        "type": "object",
        "required": [
          "data",
          "indexer_info"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "amount"
            ],
            "properties": {
              "amount": {
                "type": "string"
              },
              "id": {
                "type": "string"
              }
            }
          },
          "indexer_info": {
            "$ref": "#/components/schemas/IndexerInfo"
          }
        }
      },
...
      "ServeResponse_Vec_RuneAndAmount": {
        "type": "object",
        "required": [
          "data",
          "indexer_info"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "amount"
              ],
              "properties": {
                "amount": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                }
              }
            }
          },
          "indexer_info": {
            "$ref": "#/components/schemas/IndexerInfo"
          }
        }
      },
      "ServeResponse_Vec_RuneEdict": {
        "type": "object",
        "required": [
          "data",
          "indexer_info"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "rune_id",
                "amount",
                "block_height"
              ],
              "properties": {
                "amount": {
                  "type": "string"
                },
                "block_height": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "output": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int32",
                  "minimum": 0
                },
                "rune_id": {
                  "type": "string"
                }
              }
            }
          },
          "indexer_info": {
            "$ref": "#/components/schemas/IndexerInfo"
          }
        }
      },
      "ServeResponse_Vec_RuneUtxo": {
        "type": "object",
        "required": [
          "data",
          "indexer_info"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "tx_hash",
                "output_index",
                "height",
                "satoshis",
                "runes"
              ],
              "properties": {
                "height": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "output_index": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "runes": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuneAndAmount"
                  }
                },
                "satoshis": {
                  "type": "string"
                },
                "tx_hash": {
                  "type": "string"
                }
              }
            }
          },
          "indexer_info": {
            "$ref": "#/components/schemas/IndexerInfo"
          }
        }
      }
    }
  }
}